// types/EditProfileFormData.ts
export interface EditProfileFormData {
  nome: string;
  email: string;
  telefone: string;
  cpf: string;
  dataNascimento: string;
  pais: string;
  estado: string;
  cidade: string;
  cep: string;
  rua: string;
  numero: string;
  apartamento: string;
  logradouro: string;
  imagemPerfil?: string;
  senha?: string;
  confirmarSenha?: string;
}